+++ publish/ContractPublisher.ts
@@ -1,6 +1,5 @@
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 import { ethers } from "@nomiclabs/buidler"
-import { BigNumber, Signer } from "ethers"
 import { ExternalContracts, Layer } from "../scripts/common"
 import {
     AmmReader,
@@ -11,16 +10,12 @@ import {
     InsuranceFund,
     L2PriceFeed,
     MetaTxGateway,
-    PerpToken,
     RootBridge,
-    TetherToken,
 } from "../types/ethers"
-import { IMultiTokenMediator } from "../types/ethers/IMultiTokenMediator"
 import { ContractWrapperFactory } from "./contract/ContractWrapperFactory"
 import { DeployConfig } from "./contract/DeployConfig"
 import { AmmInstanceName, ContractName } from "./ContractName"
 import { OzContractDeployer } from "./OzContractDeployer"
-import { OzScript } from "./OzScript"
 import { SettingsDao } from "./SettingsDao"
 import { SystemMetadataDao } from "./SystemMetadataDao"
 
@@ -31,112 +26,32 @@ export class ContractPublisher {
     readonly externalContract: ExternalContracts
     readonly factory: ContractWrapperFactory
     readonly deployConfig: DeployConfig
-    private signer!: Signer
-
-    readonly taskBatchesMap: Record<Layer, DeployTask[][]> = {
+    protected taskBatchesMap: Record<Layer, DeployTask[][]> = {
         layer1: [
             // batch 0
             [
                 async (): Promise<void> => {
-                    if (!this.deployConfig.deployUsdt) {
-                        return
-                    }
-
-                    // deploy USDT
-                    const tetherContract = this.factory.create<TetherToken>(ContractName.TetherToken)
-                    await tetherContract.deployImmutableContract(
-                        this.deployConfig.usdtInitSupply,
-                        "Tether USD",
-                        "USDT",
-                        6,
-                    )
-                },
-                async (): Promise<void> => {
-                    if (!this.deployConfig.deployPerp) {
-                        return
-                    }
-
-                    const perpContract = this.factory.create<PerpToken>(ContractName.PerpToken)
-                    await perpContract.deployImmutableContract(this.deployConfig.perpInitSupply)
-                },
-                async (): Promise<void> => {
-                    console.log("distribute PERP to faucet...(should only do it on testnets)")
-                    const perpToken = await this.factory.create<PerpToken>(ContractName.PerpToken).instance()
-                    if (this.externalContract.testnetFaucet) {
-                        await perpToken.transfer(
-                            this.externalContract.testnetFaucet!,
-                            this.deployConfig.perpInitSupply.toString(),
-                        )
-                    }
-                },
-                async (): Promise<void> => {
                     // deploy root bridge
-                    const rootBridgeContract = this.factory.create<RootBridge>(ContractName.RootBridge)
-                    await rootBridgeContract.deployUpgradableContract(
-                        this.externalContract.ambBridgeOnEth!,
-                        this.externalContract.multiTokenMediatorOnEth!,
-                    )
+                    await this.factory
+                        .create<RootBridge>(ContractName.RootBridge)
+                        .deployUpgradableContract(
+                            this.externalContract.ambBridgeOnEth!,
+                            this.externalContract.multiTokenMediatorOnEth!,
+                        )
                 },
             ],
             // batch 1
             [
                 async (): Promise<void> => {
-                    // only distribute USDT in testnets
-                    if (!this.deployConfig.usdtToFaucet) {
-                        return
-                    }
-
-                    console.log("distributing USDT to insurance fund and arbitrageur...")
-
-                    // prepare address from layer 1 and layer 2
-                    const tetherOnEth = this.settingsDao.getExternalContracts("layer1").tether!
-                    const mediatorOnEth = this.settingsDao.getExternalContracts("layer1").multiTokenMediatorOnEth!
-                    const insuranceFundOnXdai = this.systemMetadataDao.getContractMetadata(
-                        "layer2",
-                        ContractName.InsuranceFund,
-                    ).address
-                    const arbOnXdai = this.settingsDao.getExternalContracts("layer2").arbitrageur!
-
-                    // prepare contract instance
-                    const tether = (await ethers.getContractAt(ContractName.TetherToken, tetherOnEth)) as TetherToken
-                    const mediator = (await ethers.getContractAt(
-                        ContractName.MultiTokenMediatorMock,
-                        mediatorOnEth,
-                    )) as IMultiTokenMediator
-                    const tetherDigits = await tether.decimals()
-
-                    // approve tether and distribute to arb and insurance fund
-                    console.log("tether.approve...")
-                    await (await tether.approve(mediatorOnEth, "0")).wait(this.confirmations)
-                    await (await tether.approve(mediatorOnEth, ethers.constants.MaxUint256.toString())).wait(
-                        this.confirmations,
-                    )
-
-                    console.log("relayTokens...")
-                    await mediator.relayTokens(
-                        tetherOnEth,
-                        insuranceFundOnXdai,
-                        BigNumber.from("100000000")
-                            .mul(tetherDigits)
-                            .toString(),
-                    )
-                    await mediator.relayTokens(
-                        tetherOnEth,
-                        arbOnXdai,
-                        BigNumber.from("100000000")
-                            .mul(tetherDigits)
-                            .toString(),
-                    )
-                },
-                async (): Promise<void> => {
-                    console.log("deploy chainlink price feed on L1")
+                    console.log("deploy chainlink price feed on layer 1...")
                     const l2PriceFeedOnXdai = this.systemMetadataDao.getContractMetadata(
                         "layer2",
                         ContractName.L2PriceFeed,
                     ).address
                     const rootBridgeContract = this.factory.create<RootBridge>(ContractName.RootBridge)
-                    const chainlinkContract = this.factory.create<ChainlinkL1>(ContractName.ChainlinkL1)
-                    await chainlinkContract.deployUpgradableContract(rootBridgeContract.address!, l2PriceFeedOnXdai)
+                    await this.factory
+                        .create<ChainlinkL1>(ContractName.ChainlinkL1)
+                        .deployUpgradableContract(rootBridgeContract.address!, l2PriceFeedOnXdai)
                 },
                 async (): Promise<void> => {
                     console.log("setPriceFeed...")
@@ -145,16 +60,28 @@ export class ContractPublisher {
                     await (await rootBridge.setPriceFeed(chainlinkContract.address!)).wait(this.confirmations)
                 },
                 async (): Promise<void> => {
-                    console.log("add aggregator of chainlink price feed on L1...")
+                    console.log("add aggregator of chainlink price feed on layer 1...")
                     const chainlinkContract = this.factory.create<ChainlinkL1>(ContractName.ChainlinkL1)
                     const chainlink = await chainlinkContract.instance()
                     for (const priceFeedKey in this.deployConfig.chainlinkMap) {
                         const address = this.deployConfig.chainlinkMap[priceFeedKey]
                         console.log(`addAggregator=${priceFeedKey.toString()}`)
-                        await (await chainlink.addAggregator(ethers.utils.id(priceFeedKey.toString()), address)).wait(
-                            this.confirmations,
-                        )
+                        await (
+                            await chainlink.addAggregator(
+                                ethers.utils.formatBytes32String(priceFeedKey.toString()),
+                                address,
+                            )
+                        ).wait(this.confirmations)
                     }
+                },
+                async (): Promise<void> => {
+                    const gov = this.externalContract.foundationGovernance!
+                    console.log(`transferring owner to governance=${gov}...please remember to claim the ownership`)
+
+                    const chainlinkL1 = await this.factory.create<ChainlinkL1>(ContractName.ChainlinkL1).instance()
+                    const rootBridge = await this.factory.create<RootBridge>(ContractName.RootBridge).instance()
+                    await (await chainlinkL1.setOwner(gov)).wait(this.confirmations)
+                    await (await rootBridge.setOwner(gov)).wait(this.confirmations)
                 },
             ],
         ],
@@ -164,68 +91,56 @@ export class ContractPublisher {
                 async (): Promise<void> => {
                     // deploy meta tx gateway
                     const chainId = this.settingsDao.getChainId("layer1")
-                    const metaTxGatewayContract = this.factory.create<MetaTxGateway>(ContractName.MetaTxGateway)
-                    await metaTxGatewayContract.deployUpgradableContract("Perp", "1", chainId)
+                    await this.factory
+                        .create<MetaTxGateway>(ContractName.MetaTxGateway)
+                        .deployUpgradableContract("Perp", "1", chainId)
                 },
                 async (): Promise<void> => {
                     // deploy client bridge
                     const ambBridgeOnXDai = this.externalContract.ambBridgeOnXDai!
                     const multiTokenMediatorOnXDai = this.externalContract.multiTokenMediatorOnXDai!
                     const metaTxGatewayContract = this.factory.create<MetaTxGateway>(ContractName.MetaTxGateway)
-                    const clientBridgeContract = this.factory.create<ClientBridge>(ContractName.ClientBridge)
-                    await clientBridgeContract.deployUpgradableContract(
-                        ambBridgeOnXDai,
-                        multiTokenMediatorOnXDai,
-                        metaTxGatewayContract.address!,
-                    )
-                },
-                async (): Promise<void> => {
-                    console.log("metaTxGateway.addToWhitelists(clientBridge.address)...")
-                    const metaTxGatewayContract = this.factory.create<MetaTxGateway>(ContractName.MetaTxGateway)
-                    const metaTxGateway = await metaTxGatewayContract.instance()
-                    const clientBridgeContract = this.factory.create<ClientBridge>(ContractName.ClientBridge)
-                    await (await metaTxGateway.addToWhitelists(clientBridgeContract.address!)).wait(this.confirmations)
+                    await this.factory
+                        .create<ClientBridge>(ContractName.ClientBridge)
+                        .deployUpgradableContract(
+                            ambBridgeOnXDai,
+                            multiTokenMediatorOnXDai,
+                            metaTxGatewayContract.address!,
+                        )
                 },
                 async (): Promise<void> => {
                     // deploy insurance fund
-                    const insuranceFundContract = this.factory.create<InsuranceFund>(ContractName.InsuranceFund)
-                    await insuranceFundContract.deployUpgradableContract()
+                    await this.factory.create<InsuranceFund>(ContractName.InsuranceFund).deployUpgradableContract()
                 },
                 async (): Promise<void> => {
                     // deploy L2 price feed
-                    const ambBridgeOnXDaiAddr = this.settingsDao.getExternalContracts("layer2").ambBridgeOnXDai!
+                    const ambBridgeOnXDaiAddr = this.externalContract.ambBridgeOnXDai!
                     const rootBridgeOnEthAddr = this.systemMetadataDao.getContractMetadata(
                         "layer1",
                         ContractName.RootBridge,
                     ).address
-                    const l2PriceFeedContract = this.factory.create<L2PriceFeed>(ContractName.L2PriceFeed)
-                    await l2PriceFeedContract.deployUpgradableContract(ambBridgeOnXDaiAddr, rootBridgeOnEthAddr)
-                },
-                async (): Promise<void> => {
-                    console.log("add aggregators to L2PriceFeed")
-                    const l2PriceFeed = await this.factory.create<L2PriceFeed>(ContractName.L2PriceFeed).instance()
-                    for (const priceFeedKey in this.deployConfig.chainlinkMap) {
-                        console.log(`add aggregator=${priceFeedKey}`)
-                        await (await l2PriceFeed.addAggregator(ethers.utils.id(priceFeedKey.toString()))).wait(
-                            this.confirmations,
-                        )
-                    }
+                    await this.factory
+                        .create<L2PriceFeed>(ContractName.L2PriceFeed)
+                        .deployUpgradableContract(ambBridgeOnXDaiAddr, rootBridgeOnEthAddr)
                 },
                 async (): Promise<void> => {
                     // deploy clearing house
                     const insuranceFundContract = this.factory.create<InsuranceFund>(ContractName.InsuranceFund)
                     const metaTxGatewayContract = this.factory.create<MetaTxGateway>(ContractName.MetaTxGateway)
-                    const clearingHouseContract = this.factory.create<ClearingHouse>(ContractName.ClearingHouse)
-                    await clearingHouseContract.deployUpgradableContract(
-                        this.deployConfig.initMarginRequirement,
-                        this.deployConfig.maintenanceMarginRequirement,
-                        this.deployConfig.liquidationFeeRatio,
-                        insuranceFundContract.address!,
-                        metaTxGatewayContract.address!,
-                    )
+                    await this.factory
+                        .create<ClearingHouse>(ContractName.ClearingHouse)
+                        .deployUpgradableContract(
+                            this.deployConfig.initMarginRequirement,
+                            this.deployConfig.maintenanceMarginRequirement,
+                            this.deployConfig.liquidationFeeRatio,
+                            insuranceFundContract.address!,
+                            metaTxGatewayContract.address!,
+                        )
                 },
                 async (): Promise<void> => {
-                    const clearingHouseContract = this.factory.create<ClearingHouse>(ContractName.ClearingHouse)
+                    const clearingHouse = await this.factory
+                        .create<ClearingHouse>(ContractName.ClearingHouse)
+                        .instance()
                     const metaTxGateway = await this.factory
                         .create<MetaTxGateway>(ContractName.MetaTxGateway)
                         .instance()
@@ -234,81 +149,35 @@ export class ContractPublisher {
                         .instance()
 
                     console.log("metaTxGateway.addToWhitelists...")
-                    await (await metaTxGateway.addToWhitelists(clearingHouseContract.address!)).wait(this.confirmations)
+                    await (await metaTxGateway.addToWhitelists(clearingHouse.address)).wait(this.confirmations)
                     console.log("insuranceFundContract.setBeneficiary...")
-                    await (await insuranceFund.setBeneficiary(clearingHouseContract.address!)).wait(this.confirmations)
+                    await (await insuranceFund.setBeneficiary(clearingHouse.address)).wait(this.confirmations)
+                    console.log("clearingHouse.setWhitelist...")
+                    await (
+                        await clearingHouse.setWhitelist(this.settingsDao.getExternalContracts("layer2").arbitrageur!)
+                    ).wait(this.confirmations)
                 },
                 async (): Promise<void> => {
                     // deploy amm
                     const l2PriceFeedContract = this.factory.create<L2PriceFeed>(ContractName.L2PriceFeed)
-                    const ammContract = this.factory.createAmm(AmmInstanceName.ETHUSDT)
-                    const tetherAddr = this.deployConfig.deployUsdt
-                        ? // USDT is custom-managed by us
-                          this.systemMetadataDao.getContractMetadata("layer1", ContractName.TetherToken).address
-                        : // USDT is externally managed
-                          this.settingsDao.getExternalContracts("layer2").tether!
+                    const ammContract = this.factory.createAmm(AmmInstanceName.ETHUSDC)
+                    const quoteTokenAddr = this.externalContract.usdc!
                     await ammContract.deployUpgradableContract(
                         this.deployConfig.ammConfigMap,
                         l2PriceFeedContract.address!,
-                        tetherAddr,
+                        quoteTokenAddr,
                     )
-                },
-                async (): Promise<void> => {
-                    // setup amm
-                    console.log("setting up Amm ETHUSDT...")
-                    const insuranceFundContract = this.factory.create<InsuranceFund>(ContractName.InsuranceFund)
-                    const clearingHouseContract = this.factory.create<ClearingHouse>(ContractName.ClearingHouse)
-                    const ammContract = this.factory.createAmm(AmmInstanceName.ETHUSDT)
-                    const amm = await ammContract.instance()
-                    const insuranceFund = await insuranceFundContract.instance()
-                    const { maxHoldingBaseAsset } = this.deployConfig.ammConfigMap[AmmInstanceName.ETHUSDT].properties
-
-                    if (maxHoldingBaseAsset.gt(0)) {
-                        console.log("setMaxHoldingBaseAsset...")
-                        await (await amm.setMaxHoldingBaseAsset({ d: maxHoldingBaseAsset.toString() })).wait(
-                            this.confirmations,
-                        )
-                    }
-                    console.log("amm.setCounterParty...")
-                    await (await amm.setCounterParty(clearingHouseContract.address!)).wait(this.confirmations)
-                    console.log("insuranceFund.addAmm...")
-                    await (await insuranceFund.addAmm(amm.address)).wait(this.confirmations)
                 },
                 async (): Promise<void> => {
                     // deploy amm
                     const l2PriceFeedContract = this.factory.create<L2PriceFeed>(ContractName.L2PriceFeed)
-                    const ammContract = this.factory.createAmm(AmmInstanceName.BTCUSDT)
-                    const tetherAddr = this.deployConfig.deployUsdt
-                        ? // USDT is custom-managed by us
-                          this.systemMetadataDao.getContractMetadata("layer1", ContractName.TetherToken).address
-                        : // USDT is externally managed
-                          this.settingsDao.getExternalContracts("layer2").tether!
+                    const ammContract = this.factory.createAmm(AmmInstanceName.BTCUSDC)
+                    const quoteTokenAddr = this.externalContract.usdc!
                     await ammContract.deployUpgradableContract(
                         this.deployConfig.ammConfigMap,
                         l2PriceFeedContract.address!,
-                        tetherAddr,
+                        quoteTokenAddr,
                     )
-                },
-                async (): Promise<void> => {
-                    // setup amm
-                    console.log("setting up Amm BTCUSDT...")
-                    const insuranceFundContract = this.factory.create<InsuranceFund>(ContractName.InsuranceFund)
-                    const clearingHouseContract = this.factory.create<ClearingHouse>(ContractName.ClearingHouse)
-                    const ammContract = this.factory.createAmm(AmmInstanceName.BTCUSDT)
-                    const amm = await ammContract.instance()
-                    const insuranceFund = await insuranceFundContract.instance()
-                    const { maxHoldingBaseAsset } = this.deployConfig.ammConfigMap[AmmInstanceName.BTCUSDT].properties
-
-                    if (maxHoldingBaseAsset.gt(0)) {
-                        console.log("setMaxHoldingBaseAsset...")
-                        await (await amm.setMaxHoldingBaseAsset({ d: maxHoldingBaseAsset.toString() })).wait(
-                            this.confirmations,
-                        )
-                    }
-                    console.log("amm.setCounterParty...")
-                    await (await amm.setCounterParty(clearingHouseContract.address!)).wait(this.confirmations)
-                    console.log("insuranceFund.addAmm...")
-                    await (await insuranceFund.addAmm(amm.address)).wait(this.confirmations)
                 },
                 async (): Promise<void> => {
                     //  deploy clearingHouseViewer
@@ -324,16 +193,85 @@ export class ContractPublisher {
                     await ammReaderContract.deployImmutableContract()
                 },
                 async (): Promise<void> => {
-                    // open amm
-                    console.log("opening Amm ETHUSDT...")
-                    const ethUsdt = await this.factory.createAmm(AmmInstanceName.ETHUSDT).instance()
-                    await (await ethUsdt.setOpen(true)).wait(this.confirmations)
+                    console.log("metaTxGateway.addToWhitelists(clientBridge.address)...")
+                    const metaTxGatewayContract = this.factory.create<MetaTxGateway>(ContractName.MetaTxGateway)
+                    const metaTxGateway = await metaTxGatewayContract.instance()
+                    const clientBridgeContract = this.factory.create<ClientBridge>(ContractName.ClientBridge)
+                    await (await metaTxGateway.addToWhitelists(clientBridgeContract.address!)).wait(this.confirmations)
+                },
+                async (): Promise<void> => {
+                    console.log("add aggregators to L2PriceFeed")
+                    const l2PriceFeed = await this.factory.create<L2PriceFeed>(ContractName.L2PriceFeed).instance()
+                    for (const priceFeedKey in this.deployConfig.chainlinkMap) {
+                        console.log(`add aggregator=${priceFeedKey}`)
+                        await (
+                            await l2PriceFeed.addAggregator(ethers.utils.formatBytes32String(priceFeedKey.toString()))
+                        ).wait(this.confirmations)
+                    }
+                },
+                async (): Promise<void> => {
+                    // setup amm
+                    console.log("setting up Amm ETHUSDC...")
+                    const insuranceFundContract = this.factory.create<InsuranceFund>(ContractName.InsuranceFund)
+                    const clearingHouseContract = this.factory.create<ClearingHouse>(ContractName.ClearingHouse)
+                    const ammContract = this.factory.createAmm(AmmInstanceName.ETHUSDC)
+                    const amm = await ammContract.instance()
+                    const insuranceFund = await insuranceFundContract.instance()
+                    const { maxHoldingBaseAsset, openInterestNotionalCap } = this.deployConfig.ammConfigMap[
+                        AmmInstanceName.ETHUSDC
+                    ].properties
+
+                    if (maxHoldingBaseAsset.gt(0)) {
+                        console.log("setCap...")
+                        await (
+                            await amm.setCap(
+                                { d: maxHoldingBaseAsset.toString() },
+                                { d: openInterestNotionalCap.toString() },
+                            )
+                        ).wait(this.confirmations)
+                    }
+                    console.log("amm.setCounterParty...")
+                    await (await amm.setCounterParty(clearingHouseContract.address!)).wait(this.confirmations)
+                    console.log("insuranceFund.addAmm...")
+                    await (await insuranceFund.addAmm(amm.address)).wait(this.confirmations)
+                },
+                async (): Promise<void> => {
+                    // setup amm
+                    console.log("setting up Amm BTCUSDC...")
+                    const insuranceFundContract = this.factory.create<InsuranceFund>(ContractName.InsuranceFund)
+                    const clearingHouseContract = this.factory.create<ClearingHouse>(ContractName.ClearingHouse)
+                    const ammContract = this.factory.createAmm(AmmInstanceName.BTCUSDC)
+                    const amm = await ammContract.instance()
+                    const insuranceFund = await insuranceFundContract.instance()
+                    const { maxHoldingBaseAsset, openInterestNotionalCap } = this.deployConfig.ammConfigMap[
+                        AmmInstanceName.BTCUSDC
+                    ].properties
+
+                    if (maxHoldingBaseAsset.gt(0)) {
+                        console.log("setCap...")
+                        await (
+                            await amm.setCap(
+                                { d: maxHoldingBaseAsset.toString() },
+                                { d: openInterestNotionalCap.toString() },
+                            )
+                        ).wait(this.confirmations)
+                    }
+                    console.log("amm.setCounterParty...")
+                    await (await amm.setCounterParty(clearingHouseContract.address!)).wait(this.confirmations)
+                    console.log("insuranceFund.addAmm...")
+                    await (await insuranceFund.addAmm(amm.address)).wait(this.confirmations)
                 },
                 async (): Promise<void> => {
                     // open amm
-                    console.log("opening Amm ETHUSDT...")
-                    const btcUsdt = await this.factory.createAmm(AmmInstanceName.BTCUSDT).instance()
-                    await (await btcUsdt.setOpen(true)).wait(this.confirmations)
+                    console.log("opening Amm ETHUSDC...")
+                    const ethUsdc = await this.factory.createAmm(AmmInstanceName.ETHUSDC).instance()
+                    await (await ethUsdc.setOpen(true)).wait(this.confirmations)
+                },
+                async (): Promise<void> => {
+                    // open amm
+                    console.log("opening Amm BTCUSDC...")
+                    const btcUsdc = await this.factory.createAmm(AmmInstanceName.BTCUSDC).instance()
+                    await (await btcUsdc.setOpen(true)).wait(this.confirmations)
                 },
                 async (): Promise<void> => {
                     const l2PriceFeed = await this.factory.create<L2PriceFeed>(ContractName.L2PriceFeed).instance()
@@ -343,6 +281,32 @@ export class ContractPublisher {
                         )
                     ).wait(this.confirmations)
                 },
+                async (): Promise<void> => {
+                    const gov = this.externalContract.foundationGovernance!
+                    console.log(`transferring owner to governance=${gov}...please remember to claim the ownership`)
+
+                    const metaTxGateway = await this.factory
+                        .create<MetaTxGateway>(ContractName.MetaTxGateway)
+                        .instance()
+                    const clientBridge = await this.factory.create<ClientBridge>(ContractName.ClientBridge).instance()
+                    const insuranceFund = await this.factory
+                        .create<InsuranceFund>(ContractName.InsuranceFund)
+                        .instance()
+                    const l2PriceFeed = await this.factory.create<L2PriceFeed>(ContractName.L2PriceFeed).instance()
+                    const clearingHouse = await this.factory
+                        .create<ClearingHouse>(ContractName.ClearingHouse)
+                        .instance()
+                    const ETHUSDC = await this.factory.createAmm(AmmInstanceName.ETHUSDC).instance()
+                    const BTCUSDC = await this.factory.createAmm(AmmInstanceName.BTCUSDC).instance()
+
+                    // await (await metaTxGateway.setOwner(gov)).wait(this.confirmations)
+                    // await (await clientBridge.setOwner(gov)).wait(this.confirmations)
+                    // await (await insuranceFund.setOwner(gov)).wait(this.confirmations)
+                    // await (await l2PriceFeed.setOwner(gov)).wait(this.confirmations)
+                    // await (await clearingHouse.setOwner(gov)).wait(this.confirmations)
+                    // await (await ETHUSDC.setOwner(gov)).wait(this.confirmations)
+                    await (await BTCUSDC.setOwner(gov)).wait(this.confirmations)
+                },
             ],
         ],
     }
@@ -351,7 +315,6 @@ export class ContractPublisher {
         readonly layerType: Layer,
         readonly settingsDao: SettingsDao,
         readonly systemMetadataDao: SystemMetadataDao,
-        readonly ozScript: OzScript,
     ) {
         this.externalContract = settingsDao.getExternalContracts(layerType)
         this.deployConfig = new DeployConfig(settingsDao.stage)
@@ -399,15 +362,13 @@ export class ContractPublisher {
             return
         }
         // local are basically in 1 layer, can't transfer twice in the same network. will transfer in the very last batch
-        if (this.settingsDao.isLocal()) {
+        if (this.settingsDao.getChainId("layer1") === this.settingsDao.getChainId("layer2")) {
             const layerWithMoreBatch =
                 this.taskBatchesMap.layer1.length > this.taskBatchesMap.layer2.length ? "layer1" : "layer2"
             if (layerWithMoreBatch !== this.layerType) {
                 return
             }
         }
-
-        // TODO add a prompt
         const governance = this.externalContract.foundationGovernance!
         console.log(`${this.layerType} batch ends, transfer proxy admin to ${governance}`)
         await OzContractDeployer.transferProxyAdminOwnership(governance)