+++ publish/OzContractDeployer.ts
@@ -22,7 +22,6 @@ export class OzContractDeployer {
 
     async deploy(contractFileName: string, args: any[]): Promise<string> {
         // deploy contract by open zeppelin upgrade plugin
-        // ozScript won't replace the existing one, we have to manually remove it before deploy new contract first time
         const contract = await ethers.getContractFactory(contractFileName)
         const instance = await upgrades.deployProxy(contract, args, {
             initializer: this.ozInitMethodName,
@@ -34,7 +33,10 @@ export class OzContractDeployer {
 
     async prepareUpgrade(proxy: string, contractFileName: string): Promise<string> {
         const factory = await ethers.getContractFactory(contractFileName)
-        const address = await upgrades.prepareUpgrade(proxy, factory)
+        const address = await upgrades.prepareUpgrade(proxy, factory, {
+            unsafeAllowCustomTypes: true,
+        })
+        console.log(`prepareUpgrade proxy=${proxy}, contractFileName=${contractFileName}, address=${address}`)
         const contract = await ethers.getContractAt(contractFileName, address)
         await this.syncNonce(contract.deployTransaction.hash)
         return address