+++ publish/contract/DeployConfig.ts
@@ -2,13 +2,13 @@ import { BigNumber } from "ethers"
 import { Stage } from "../../scripts/common"
 import { AmmInstanceName } from "../ContractName"
 
-const DEFAULT_DIGITS = BigNumber.from("1000000000000000000")
+// TODO replace by ethers format
+const DEFAULT_DIGITS = BigNumber.from(10).pow(18)
 
 // chainlink
 export enum PriceFeedKey {
     BTC = "BTC",
     ETH = "ETH",
-    LINK = "LINK",
 }
 
 // amm
@@ -25,19 +25,23 @@ interface AmmDeployArgs {
 
 interface AmmProperties {
     maxHoldingBaseAsset: BigNumber
+    openInterestNotionalCap: BigNumber
 }
 
 export type AmmConfig = { deployArgs: AmmDeployArgs; properties: AmmProperties }
 export type AmmConfigMap = Record<string, AmmConfig>
 export const BTC_USD_AMM: AmmConfig = {
     deployArgs: {
-        quoteAssetReserve: BigNumber.from(13095000 * 2).mul(DEFAULT_DIGITS),
-        baseAssetReserve: BigNumber.from(1000 * 2).mul(DEFAULT_DIGITS),
+        // base * price
+        quoteAssetReserve: BigNumber.from(10000000).mul(DEFAULT_DIGITS),
+        baseAssetReserve: BigNumber.from(100).mul(DEFAULT_DIGITS), // 100 BTC
         tradeLimitRatio: BigNumber.from(90)
             .mul(DEFAULT_DIGITS)
             .div(100), // 90% trading limit ratio
         fundingPeriod: BigNumber.from(3600), // 1 hour
-        fluctuation: BigNumber.from(0),
+        fluctuation: BigNumber.from(12)
+            .mul(DEFAULT_DIGITS)
+            .div(1000), // 1.2%
         priceFeedKey: PriceFeedKey.BTC,
         tollRatio: BigNumber.from(0)
             .mul(DEFAULT_DIGITS)
@@ -47,19 +51,25 @@ export const BTC_USD_AMM: AmmConfig = {
             .div(10000), // 0.1%
     },
     properties: {
-        maxHoldingBaseAsset: BigNumber.from(0),
+        maxHoldingBaseAsset: BigNumber.from(DEFAULT_DIGITS)
+            .mul(5)
+            .div(1000), // 0.05 BTC ~= $1000 USD,
+        openInterestNotionalCap: BigNumber.from(DEFAULT_DIGITS).mul(100000), // $100K
     },
 }
 
 export const ETH_USD_AMM: AmmConfig = {
     deployArgs: {
-        quoteAssetReserve: BigNumber.from(40958000).mul(DEFAULT_DIGITS),
-        baseAssetReserve: BigNumber.from(100000).mul(DEFAULT_DIGITS),
+        // base * price
+        quoteAssetReserve: BigNumber.from(5000000).mul(DEFAULT_DIGITS),
+        baseAssetReserve: BigNumber.from(2000).mul(DEFAULT_DIGITS), // 2000 ETH
         tradeLimitRatio: BigNumber.from(90)
             .mul(DEFAULT_DIGITS)
             .div(100), // 90% trading limit ratio
         fundingPeriod: BigNumber.from(3600), // 1 hour
-        fluctuation: BigNumber.from(0),
+        fluctuation: BigNumber.from(12)
+            .mul(DEFAULT_DIGITS)
+            .div(1000), // 1.2%
         priceFeedKey: PriceFeedKey.ETH,
         tollRatio: BigNumber.from(0)
             .mul(DEFAULT_DIGITS)
@@ -69,22 +79,14 @@ export const ETH_USD_AMM: AmmConfig = {
             .div(10000), // 0.1%
     },
     properties: {
-        maxHoldingBaseAsset: BigNumber.from(0),
+        maxHoldingBaseAsset: BigNumber.from(2), // 2 ETH ~= $1000 USD
+        openInterestNotionalCap: BigNumber.from(DEFAULT_DIGITS).mul(100000), // $100K
     },
 }
 
 export class DeployConfig {
     // deploy
     readonly confirmations: number
-
-    // token
-    readonly deployUsdt: boolean
-    readonly usdtToFaucet: boolean
-    readonly usdtInitSupply: BigNumber = BigNumber.from("10000000000").mul("1000000")
-
-    // perp
-    readonly deployPerp: boolean
-    readonly perpInitSupply = BigNumber.from("1500000000").mul(DEFAULT_DIGITS)
 
     // chainlink
     readonly chainlinkMap: Record<string, string>
@@ -102,44 +104,32 @@ export class DeployConfig {
 
     // amm
     readonly ammConfigMap: Record<string, AmmConfig> = {
-        [AmmInstanceName.BTCUSDT]: BTC_USD_AMM,
-        [AmmInstanceName.ETHUSDT]: ETH_USD_AMM,
+        [AmmInstanceName.BTCUSDC]: BTC_USD_AMM,
+        [AmmInstanceName.ETHUSDC]: ETH_USD_AMM,
     }
 
     constructor(stage: Stage) {
         switch (stage) {
             case "production":
                 this.confirmations = 5
-                this.deployUsdt = false
-                this.deployPerp = false
-                this.usdtToFaucet = false
                 this.chainlinkMap = {
                     [PriceFeedKey.BTC]: "0xF4030086522a5bEEa4988F8cA5B36dbC97BeE88c",
                     [PriceFeedKey.ETH]: "0x5f4eC3Df9cbd43714FE2740f5E3616155c5b8419",
-                    [PriceFeedKey.LINK]: "0x2c1d072e956AFFC0D435Cb7AC38EF18d24d9127c",
                 }
                 break
             case "staging":
                 this.confirmations = 5
-                this.deployUsdt = false
-                this.deployPerp = true
-                this.usdtToFaucet = false
                 this.chainlinkMap = {
                     [PriceFeedKey.BTC]: "0xECe365B379E1dD183B20fc5f022230C044d51404",
                     [PriceFeedKey.ETH]: "0x8A753747A1Fa494EC906cE90E9f37563A8AF630e",
-                    [PriceFeedKey.LINK]: "0xd8bD0a1cB028a31AA859A21A3758685a95dE4623",
                 }
                 break
             case "test":
                 this.confirmations = 1
-                this.deployUsdt = true
-                this.deployPerp = true
-                this.usdtToFaucet = false
                 this.chainlinkMap = {
                     // fake address
                     [PriceFeedKey.BTC]: "0xECe365B379E1dD183B20fc5f022230C044d51404",
                     [PriceFeedKey.ETH]: "0x8A753747A1Fa494EC906cE90E9f37563A8AF630e",
-                    [PriceFeedKey.LINK]: "0xd8bD0a1cB028a31AA859A21A3758685a95dE4623",
                 }
                 break
             default: