+++ scripts/deploy-layer.ts
@@ -2,12 +2,9 @@
 import { ethers } from "@nomiclabs/buidler"
 import { BuidlerRuntimeEnvironment } from "@nomiclabs/buidler/types"
 import { ContractPublisher } from "../publish/ContractPublisher"
-import { OzScript } from "../publish/OzScript"
 import { SettingsDao } from "../publish/SettingsDao"
 import { SystemMetadataDao } from "../publish/SystemMetadataDao"
 import { AccountMetadata, Layer, Network, Stage } from "./common"
-
-export type DeployTask = () => Promise<void>
 
 export async function deployLayer(
     stage: Stage,
@@ -27,8 +24,8 @@ export async function deployLayer(
 
     const signers = await ethers.getSigners()
     const address = await signers[0].getAddress()
-    const ozScript = new OzScript(bre.web3.currentProvider, address)
-    const publisher = new ContractPublisher(layerType, settingsDao, systemMetadataDao, ozScript)
+    console.log(`deployer=${address}`)
+    const publisher = new ContractPublisher(layerType, settingsDao, systemMetadataDao)
 
     await publisher.publishContracts(batch)
 }