+++ tests/integration/clearingHouse/ClearingHouseTest.ts
@@ -81,7 +81,7 @@ describe("ClearingHouse Test", () => {
         await quoteToken.transfer(bob, toFullDigit(5000, +(await quoteToken.decimals())))
         await quoteToken.transfer(insuranceFund.address, toFullDigit(5000, +(await quoteToken.decimals())))
 
-        await amm.setMaxHoldingBaseAsset(toDecimal(0))
+        await amm.setCap(toDecimal(0), toDecimal(0))
     })
 
     async function gotoNextFundingTime(): Promise<void> {
@@ -140,6 +140,132 @@ describe("ClearingHouse Test", () => {
             const alicePosition = await clearingHouseViewer.getPersonalPositionWithFundingPayment(amm.address, alice)
             expect(alicePosition.size).to.eq(toFullDigit(-150))
             expect(alicePosition.margin).to.eq(toFullDigit(0))
+        })
+    })
+
+    describe("openInterestNotional", () => {
+        beforeEach(async () => {
+            await amm.setCap(toDecimal(0), toDecimal(600))
+            await approve(alice, clearingHouse.address, 600)
+            await approve(bob, clearingHouse.address, 600)
+        })
+
+        it("increase when increase position", async () => {
+            await clearingHouse.openPosition(amm.address, Side.BUY, toDecimal(600), toDecimal(1), toDecimal(0), {
+                from: alice,
+            })
+            expect(await clearingHouse.openInterestNotionalMap(amm.address)).eq(toFullDigitStr(600))
+        })
+
+        it("reduce when reduce position", async () => {
+            await clearingHouse.openPosition(amm.address, Side.BUY, toDecimal(600), toDecimal(1), toDecimal(0), {
+                from: alice,
+            })
+            await clearingHouse.openPosition(amm.address, Side.SELL, toDecimal(300), toDecimal(1), toDecimal(0), {
+                from: alice,
+            })
+            expect(await clearingHouse.openInterestNotionalMap(amm.address)).eq(toFullDigitStr(300))
+        })
+
+        it("reduce when close position", async () => {
+            await clearingHouse.openPosition(amm.address, Side.BUY, toDecimal(400), toDecimal(1), toDecimal(0), {
+                from: alice,
+            })
+            await clearingHouse.closePosition(amm.address, toDecimal(0), { from: alice })
+
+            // expect the result will be almost 0 (with a few rounding error)
+            const openInterestNotional = await clearingHouse.openInterestNotionalMap(amm.address)
+            expect(openInterestNotional.toNumber()).lte(10)
+        })
+
+        it("increase when traders open positions in different direction", async () => {
+            await approve(alice, clearingHouse.address, 300)
+            await clearingHouse.openPosition(amm.address, Side.BUY, toDecimal(300), toDecimal(1), toDecimal(0), {
+                from: alice,
+            })
+            await approve(bob, clearingHouse.address, 300)
+            await clearingHouse.openPosition(amm.address, Side.SELL, toDecimal(300), toDecimal(1), toDecimal(0), {
+                from: bob,
+            })
+            expect(await clearingHouse.openInterestNotionalMap(amm.address)).eq(toFullDigitStr(600))
+        })
+
+        it("increase when traders open larger position in reverse direction", async () => {
+            await approve(alice, clearingHouse.address, 600)
+            await clearingHouse.openPosition(amm.address, Side.BUY, toDecimal(250), toDecimal(1), toDecimal(0), {
+                from: alice,
+            })
+            await clearingHouse.openPosition(amm.address, Side.SELL, toDecimal(450), toDecimal(1), toDecimal(0), {
+                from: alice,
+            })
+            expect(await clearingHouse.openInterestNotionalMap(amm.address)).eq(toFullDigitStr(200))
+        })
+
+        it("is 0 when everyone close position", async () => {
+            await clearingHouse.openPosition(amm.address, Side.BUY, toDecimal(250), toDecimal(1), toDecimal(0), {
+                from: alice,
+            })
+            await clearingHouse.openPosition(amm.address, Side.SELL, toDecimal(250), toDecimal(1), toDecimal(0), {
+                from: bob,
+            })
+            await clearingHouse.closePosition(amm.address, toDecimal(0), { from: alice })
+            await clearingHouse.closePosition(amm.address, toDecimal(0), { from: bob })
+
+            // expect the result will be almost 0 (with a few rounding error)
+            const openInterestNotional = await clearingHouse.openInterestNotionalMap(amm.address)
+            expect(openInterestNotional.toNumber()).lte(10)
+        })
+
+        it("is 0 when everyone close position, one of them is bankrupt position", async () => {
+            await clearingHouse.openPosition(amm.address, Side.SELL, toDecimal(250), toDecimal(1), toDecimal(0), {
+                from: alice,
+            })
+            await clearingHouse.openPosition(amm.address, Side.BUY, toDecimal(250), toDecimal(1), toDecimal(0), {
+                from: bob,
+            })
+
+            // when alice close, it create bad debt (bob's position is bankrupt)
+            await clearingHouse.closePosition(amm.address, toDecimal(0), { from: alice })
+
+            // bypass the restrict mode
+            await forwardBlockTimestamp(15)
+            await clearingHouse.closePosition(amm.address, toDecimal(0), { from: bob })
+
+            // expect the result will be almost 0 (with a few rounding error)
+            const openInterestNotional = await clearingHouse.openInterestNotionalMap(amm.address)
+            expect(openInterestNotional.toNumber()).lte(10)
+        })
+
+        it("stop trading if it's over openInterestCap", async () => {
+            await clearingHouse.openPosition(amm.address, Side.BUY, toDecimal(600), toDecimal(1), toDecimal(0), {
+                from: alice,
+            })
+            await expectRevert(
+                clearingHouse.openPosition(amm.address, Side.BUY, toDecimal(1), toDecimal(1), toDecimal(0), {
+                    from: alice,
+                }),
+                "over limit",
+            )
+        })
+
+        it("won't be limited by the open interest cap if the trader is the whitelist", async () => {
+            await approve(alice, clearingHouse.address, 700)
+            await clearingHouse.setWhitelist(alice)
+            await clearingHouse.openPosition(amm.address, Side.BUY, toDecimal(700), toDecimal(1), toDecimal(0), {
+                from: alice,
+            })
+            expect(await clearingHouse.openInterestNotionalMap(amm.address)).eq(toFullDigitStr(700))
+        })
+
+        it("won't stop trading if it's reducing position, even it's more than cap", async () => {
+            await clearingHouse.openPosition(amm.address, Side.BUY, toDecimal(600), toDecimal(1), toDecimal(0), {
+                from: alice,
+            })
+            await amm.setCap(toDecimal(0), toDecimal(300))
+            await clearingHouse.openPosition(amm.address, Side.SELL, toDecimal(300), toDecimal(1), toDecimal(0), {
+                from: alice,
+            })
+            expect(await clearingHouse.openInterestNotionalMap(amm.address)).eq(toFullDigitStr(300))
         })
     })
 
@@ -341,7 +467,7 @@ describe("ClearingHouse Test", () => {
 
         it("add margin", async () => {
             const receipt = await clearingHouse.addMargin(amm.address, toDecimal(80), { from: alice })
-            await expectEvent.inTransaction(receipt.tx, clearingHouse, "MarginAdded", {
+            await expectEvent.inTransaction(receipt.tx, clearingHouse, "MarginChanged", {
                 sender: alice,
                 amm: amm.address,
                 amount: toFullDigit(80),
@@ -358,16 +484,14 @@ describe("ClearingHouse Test", () => {
         })
 
         it("remove margin", async () => {
-            const removedMargin = 20
-
             // remove margin 20
-            const receipt = await clearingHouse.removeMargin(amm.address, toDecimal(removedMargin), {
+            const receipt = await clearingHouse.removeMargin(amm.address, toDecimal(20), {
                 from: alice,
             })
-            await expectEvent.inTransaction(receipt.tx, clearingHouse, "MarginRemoved", {
+            await expectEvent.inTransaction(receipt.tx, clearingHouse, "MarginChanged", {
                 sender: alice,
                 amm: amm.address,
-                amount: toFullDigit(removedMargin),
+                amount: toFullDigit(-20),
             })
             await expectEvent.inTransaction(receipt.tx, quoteToken, "Transfer", {
                 from: clearingHouse.address,